+++ app/admin/user.rb
@@ -18,7 +18,6 @@ ActiveAdmin.register User do
       f.input :email
       f.input :name
       f.input :admin
-      f.input :free
     end
     f.actions
   end