+++ db/migrate/20210122225804_remove_free_from_users.rb
@@ -1,0 +1,9 @@
+class RemoveFreeFromUsers < ActiveRecord::Migration[6.0]
+  def up
+    remove_column :users, :free
+  end
+
+  def down
+    add_column :users, :free, :boolean, null: false, default: false
+  end
+end