+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed 
-Nothing yet
+- Removed distinction between "free" and "paid" users. Users have never had to pay.
 ### Fixed
 - Rails env is not unconditionally set to production by the devise initializer anymore.